{{ cookiecutter.package_name }}-project/pyproject.toml
# ==============================================================================
# The [project] table defined by Official python.org
#
# Read: https://packaging.python.org/en/latest/guides/writing-pyproject-toml/
# ==============================================================================
[project]
name = "{{ cookiecutter.package_name }}"
# Increment version before each release - follow `semantic versioning <https://semver.org/>`_
# Currently, poetry 2.1.X doesn't support dynamic versioning
# (Read https://packaging.python.org/en/latest/guides/writing-pyproject-toml/#version)
# So this value has to be aligned with the one in ``{{ cookiecutter.package_name }}/_version.py``
version = "{{ cookiecutter.version }}"
description = "An example project generated by cookiecutter-pywf_internal_proprietary."
# Read https://dev-exp-share.readthedocs.io/en/latest/search.html?q=Pick+An+Open+Source+License+For+Python+Project&check_keywords=yes&area=default
# To pick a license and update the ``license``, ``classifier`` field in ``pyproject.toml``
# And also update the ``LICENSE.txt`` file in the git repo.
license = "{{ cookiecutter.license }}"
license-files = ["LICENSE.txt", "AUTHORS.rst"]
authors = [
    { name = "{{ cookiecutter.author }}", email = "{{ cookiecutter.author_email }}" },
]
maintainers = [
    { name = "{{ cookiecutter.author }}", email = "{{ cookiecutter.author_email }}" },
]
keywords = []
readme = "README.rst"
requires-python = ">=3.10,<4.0"
# Full list of classifiers: https://pypi.org/classifiers/
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Natural Language :: English",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: MacOS",
    "Operating System :: Unix",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
]

# ------------------------------------------------------------------------------
# Core Dependencies
# ------------------------------------------------------------------------------
dependencies = [
    # A private package for testing, you may replace it with your own package
    "esc-aws-org-struct>=0.1.3,<1.0.0",
]

# ------------------------------------------------------------------------------
# Optional dependency that can be used in ``pip install ${your_project_name}[${feature_name}]``
# Sometime this is also called "extras"
# Read: https://packaging.python.org/en/latest/guides/writing-pyproject-toml/#dependencies-optional-dependencies
#
# IMPORTANT: all optional dependencies has to be compatible with the "requires-python" field
# ------------------------------------------------------------------------------
[project.optional-dependencies]

# ------------------------------------------------------------------------------
# Local Development dependenceies
# ------------------------------------------------------------------------------
dev = [
    "rich>=13.8.1,<14.0.0", # pretty print
]

# ------------------------------------------------------------------------------
# (Unit/Coverage/Integration/Load) Test dependenceies
# ------------------------------------------------------------------------------
test = [
    "pytest>=8.2.2,<9.0.0", # Testing framework
    "pytest-cov>=6.0.0,<7.0.0", # Coverage reporting
]

# ------------------------------------------------------------------------------
# Documentation build dependenceies
# ------------------------------------------------------------------------------
doc = [
    "Sphinx>=7.4.7,<8.0.0",
    "sphinx-copybutton>=0.5.2,<1.0.0", # add copy button to code block
    "sphinx-design>=0.6.1,<1.0.0", # add additional design pattern to sphinx
    "sphinx-jinja>=2.0.2,<3.0.0", # enable jinja syntax in reStructuredText
    "furo==2024.8.6", # the furo sphinx theme
    "pygments>=2.18.0,<3.0.0", # syntax highlight
    "ipython>=8.18.1,<8.19.0", # interactive Python
    "nbsphinx>=0.8.12,<1.0.0", # add jupyter notebook in sphinx doc
    "rstobj==2.0.0", # generate reStructuredText from Python code
    "docfly==3.0.3", # automaticall generate .. toctree directives and API reference doc
]

# ------------------------------------------------------------------------------
# Mise task runner dependencies
# ------------------------------------------------------------------------------
mise = [
    "PyGithub>=2.8.0,<3.0.0", # GitHub API client
    "httpx>=0.28.0,<1.0.0", # HTTP client for ReadTheDocs API
    "tomli>=2.0.0,<3.0.0; python_version < '3.11'", # TOML parser for Python 3.10
]

# Quick Links
[project.urls]
Homepage = "https://github.com/{{ cookiecutter.github_username }}/{{ cookiecutter.package_name }}-project"
Documentation = "https://{{ cookiecutter.package_name | slugify }}.pages.dev"
Repository = "https://github.com/{{ cookiecutter.github_username }}/{{ cookiecutter.package_name }}-project"
Issues = "https://github.com/{{ cookiecutter.github_username }}/{{ cookiecutter.package_name }}-project/issues"
Changelog = "https://github.com/{{ cookiecutter.github_username }}/{{ cookiecutter.package_name }}-project/blob/main/release-history.rst"
Download = "https://pypi.org/pypi/{{ cookiecutter.package_name | slugify }}#files"

# For command line interface, read: https://packaging.python.org/en/latest/guides/writing-pyproject-toml/#creating-executable-scripts
[project.scripts]

# By default, uv searches custom indexes before PyPI (custom indexes have higher priority).
# Setting "explicit = true" prevents uv from automatically searching this index for every package.
# Instead, only packages explicitly listed under [tool.uv.sources] will be resolved from this index.
# This ensures all public packages are fetched from PyPI, and only private packages go to CodeArtifact.
# Ref: https://docs.astral.sh/uv/concepts/indexes/#explicit-indexes
[[tool.uv.index]]
name = "{{ cookiecutter.aws_codeartifact_domain }}"
url = "https://{{ cookiecutter.aws_codeartifact_domain }}-{{ cookiecutter.aws_account_id }}.d.codeartifact.{{ cookiecutter.aws_region }}.amazonaws.com/pypi/{{ cookiecutter.aws_codeartifact_repository }}/simple/"
authenticate = "always"
explicit = true

# Pin private packages to the "esc" CodeArtifact index.
# Any package not listed here will be resolved from PyPI (the implicit default index).
# Ref: https://docs.astral.sh/uv/concepts/indexes/#pinning-a-package-to-an-index
# Example:
[tool.uv.sources]
esc-aws-org-struct = { index = "esc" }

[tool.setuptools.packages.find]
where = ["."]
include = [
    "{{ cookiecutter.package_name }}*"
]
exclude = [
    "{{ cookiecutter.package_name }}.tests",
    "{{ cookiecutter.package_name }}.tests.*",
    "{{ cookiecutter.package_name }}.docs",
    "{{ cookiecutter.package_name }}.docs.*",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"
